Refactor FSM and mock into extensible packages with extra tests
