+++ vaidcord/src/vaidcord/mock/__init__.py
@@ -1,0 +1,18 @@
+from .bot import MockBot
+from .builders import MockResponseBuilder, create_mock_event, create_mock_message
+from .config import MockSettings
+from .gateway import MockGateway
+from .http import MockHTTPClient
+from .types import MockEvent, MockHTTPResponse
+
+__all__ = [
+    "MockSettings",
+    "MockEvent",
+    "MockHTTPResponse",
+    "MockGateway",
+    "MockHTTPClient",
+    "MockBot",
+    "create_mock_message",
+    "create_mock_event",
+    "MockResponseBuilder",
+]
